+++ src/anemoi/inference/commands/run.py
@@ -22,12 +22,16 @@ class RunCmd(Command):
     """Inspect the contents of a checkpoint file."""
 
     def add_arguments(self, command_parser):
+        command_parser.add_argument("--defaults", action="append", help="Sources of default values.")
         command_parser.add_argument("config", help="Path to config file.")
         command_parser.add_argument("overrides", nargs="*", help="Overrides.")
 
     def run(self, args):
 
-        config = load_config(args.config, args.overrides)
+        config = load_config(args.config, args.overrides, defaults=args.defaults)
+
+        if config.description is not None:
+            LOG.info("%s", config.description)
 
         runner = DefaultRunner(config)